16210 FIX folder_config: Prevent unintentional deletion of a non-empty folder
Prior to this Werk, the REST API could delete non-empty folders without any further check. With this Werk, the endpoint now accepts the delete_method query parameter. The possible values are:
- recursive: Deletes the folder and all the elements it contains.
- abort_on_nonempty: Deletes the folder only if it is not empty

If no delete_method is provided, "recursive" is asumed and the behaviour is the same as before this Werk.

The enpoint will return a 409 status code when trying to delete a folder that contains hosts, rules, subfolders or is referenced by another object.

Use example:
`curl -X 'DELETE' 'http://example.com/my_site/check_mk/api/1.0/objects/folder_config/my_folder?delete_metod=abort_on_nonempty' -H 'accept: */*`

SUP-16540

+++ cmk/gui/openapi/endpoints/folder_config/__init__.py
@@ -44,6 +44,7 @@ from cmk.gui.logged_in import user
 from cmk.gui.openapi.endpoints.folder_config.request_schemas import (
     BulkUpdateFolder,
     CreateFolder,
+    DeleteModeField,
     EXISTING_FOLDER_PATTERN,
     MoveFolder,
     UpdateFolder,
@@ -74,6 +75,8 @@ PATH_FOLDER_FIELD = {
         pattern=EXISTING_FOLDER_PATTERN,
     )
 }
+
+FORCE_PARAM = {DeleteModeField.field_name(): DeleteModeField()}
 
 RW_PERMISSIONS = permissions.AllPerm(
     [
@@ -254,13 +257,15 @@ def bulk_update(params: Mapping[str, Any]) -> Response:
     ".../delete",
     method="delete",
     path_params=[PATH_FOLDER_FIELD],
+    query_params=[FORCE_PARAM],
     output_empty=True,
     permissions_required=RW_PERMISSIONS,
-    additional_status_codes=[401],
+    additional_status_codes=[401, 409],
 )
 def delete(params: Mapping[str, Any]) -> Response:
     """Delete a folder"""
     user.need_permission("wato.edit")
+
     folder: Folder = params["folder"]
     if (parent := folder.parent()) is None:
         raise ProblemException(
@@ -268,7 +273,16 @@ def delete(params: Mapping[str, Any]) -> Response:
             detail="Deleting the root folder is not permitted.",
             status=401,
         )
+
+    if (params["delete_mode"] != "recursive") and (not folder.is_empty() or folder.is_referenced()):
+        raise ProblemException(
+            title="Problem deleting folder.",
+            detail="Folder is not empty or is referenced by another object. Use the force parameter to delete it.",
+            status=409,
+        )
+
     parent.delete_subfolder(folder.name())
+
     return Response(status=204)
